+++ booking_scheduler_test.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 
 from booking_scheduler import BookingScheduler
 from schedule import Customer, Schedule
-from testable_sms_sender import TestableSmsSender
+from testable_sender import TestableSmsSender, TestableMailSender
 
 NOT_ON_THE_HOUR = datetime.strptime('2021/03/26 09:05', '%Y/%m/%d %H:%M')
 ON_THE_HOUR = datetime.strptime('2021/03/26 09:00', '%Y/%m/%d %H:%M')
@@ -58,7 +58,11 @@ class BookingSchedulerTest(unittest.TestCase):
         self.assertTrue(self.testable_sms_sender.is_send_method_is_called())
 
     def test_이메일이_없는_경우에는_이메일_미발송(self):
-        pass
+        testable_mail_sender = TestableMailSender()
+        schedule = Schedule(ON_THE_HOUR, UNDER_CAPACITY, CUSTOMER)
+        self.booking_scheduler.set_mail_sender(testable_mail_sender)
+        self.booking_scheduler.add_schedule(schedule)
+        self.assertEqual(testable_mail_sender.get_count_send_mail_is_called(), 0)
 
     def test_이메일이_있는_경우에는_이메일_발송(self):
         pass
